fixing BLE notifications
This is an incredibly ugly workaround for a Maven/TinyB/JNI issue.
JNI in TinyB needs the tinyb jar to be in the classpath for FindClass
to work. Unfortunately, maven does not use the classpath, but its own
loader, so JNI ends up throwing a class not found exception.
Workaround is to start the loader manually and add the jar to the
classpath.