fixing BLE startup
TinyB was not working because it needs "Specification-Version" set in
the JAR manifest file. We have the class files in the main jar, but
not this info, since maven is not adding it while creating the package.
Thus, we use TinyB from its own JAR, which has a correct manifest.